firmware: code cleanup
Removes battery voltage parameter as it was unused.
Adds err parameter for the current control error.
Inverts motor direction as it was previously incorrect.
Fixes some control calculation errors.

In its current state, the control loop does not perform well enough. The 8-bit resolution is too poor
and normalizing the signals has the drawback of not being able to add any gain to them. Possible
solutions:
 - Add gain parameter that is applied at the end of the loop
 - Normalize signalss to another range, for example [0-10] instead of [0-1]