uart: handler: Implement basic terminal-like behaviour
With the majority of the interaction logic moving to the rc car, a basic
terminal-like interface is needed. Adds basic writing/erasing/navigating
functionality to the terminal.

Adds new uart_send function which does not take any size argument, the function
auto-terminates on zero-element without sending the zero. Remanes old uart_send
to uart_send_sz.

uart_buf was becoming a bit of a misleading name for the current
implementation. Remove it and replace it with a new struct called command_line.
Add a status field to show if there is currently a command being handled. This
field is meant to be used for command-line locking commands, such as graphers.