pid: Modify PID implementation to use structs
In order to be able to control the vehicle speed as well as the servo angle
using PID regultors, multiple PID instances are required. Modify the PID
implementation so that it uses a PID struct instead of static variables. This
makes it possible to calculate the distance and servo PIDs separately.

Also cleans up some weirdness in the PID code